+++ CHANGELOG.md
@@ -15,11 +15,21 @@ x.x.x Release notes (yyyy-MM-dd)
   cannot exist independently from a `Realm`.
 * Aggregate operations are now available on `List`: `min`, `max`, `sum`,
   `average`.
-* Add missing KVO handling for moving and exchanging objects in `RLMArray` and
-  `List`.
 * Committing write transactions (via `commitWrite` / `commitWriteTransaction` and
   `write` / `transactionWithBlock`) now optionally allow for handling errors when
   the disk is out of space.
+
+### Bugfixes
+
+* None.
+
+0.95.1 Release notes (2015-09-23)
+=============================================================
+
+### Enhancements
+
+* Add missing KVO handling for moving and exchanging objects in `RLMArray` and
+  `List`.
 
 ### Bugfixes